requirements.txt
certifi==2020.12.5
chardet==4.0.0
idna==2.10
numpy==1.19.5
python-kucoin==2.1.2
requests==2.25.1
urllib3==1.26.4
websockets==9.0.1
